+++ app-main/src/main/java/net/causw/app/main/domain/notification/notification/service/implementation/NotificationPushSender.java
@@ -7,7 +7,6 @@ import java.util.Set;
 import org.springframework.stereotype.Component;
 
 import net.causw.app.main.domain.user.account.entity.user.User;
-import net.causw.app.main.shared.infra.firebase.FcmUtils;
 import net.causw.app.main.shared.infra.push.PushNotificationSender;
 
 import com.google.firebase.messaging.FirebaseMessagingException;
@@ -21,7 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 public class NotificationPushSender {
 
 	private final PushNotificationSender pushNotificationSender;
-	private final FcmUtils fcmUtils;
+	private final UserPushTokenWriter userPushTokenWriter;
 
 	/**
 	 * 유저의 모든 FCM 토큰으로 푸시 알림을 전송합니다.
@@ -33,7 +32,7 @@ public class NotificationPushSender {
 	 * @throws Exception 기타 예외 발생 시 예외 발생
 	 */
 	public void sendToUser(User user, String title, String body) {
-		fcmUtils.cleanInvalidFcmTokens(user);
+		userPushTokenWriter.cleanInvalidFcmTokens(user);
 		Set<String> tokens = new HashSet<>(user.getFcmTokens());
 		tokens.forEach(token -> trySend(user, token, title, body));
 	}
@@ -56,7 +55,7 @@ public class NotificationPushSender {
 			pushNotificationSender.send(token, title, body);
 		} catch (FirebaseMessagingException e) {
 			log.error("FCM 전송 실패: {}, 이유: {}", token, e.getMessage());
-			fcmUtils.removeFcmToken(user, token);
+			userPushTokenWriter.removeFcmToken(user, token);
 			log.info("오류 발생으로 FCM 토큰 제거됨: {}", token);
 		} catch (Exception e) {
 			log.error("FCM 전송 중 알 수 없는 예외 발생: {}", e.getMessage(), e);
