+++ app-main/src/main/java/net/causw/app/main/domain/notification/notification/service/implementation/UserPushTokenWriter.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.user.account.service.implementation;
+package net.causw.app.main.domain.notification.notification.service.implementation;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -91,5 +91,22 @@ public class UserPushTokenWriter {
 				redisUtils.deleteFcmTokenData(fcmToken);
 			}
 		}
+	}
+
+	/**
+	 * 사용자의 모든 FCM 토큰을 삭제합니다.
+	 * <p>
+	 * 회원 탈퇴 등 사용자의 모든 디바이스 세션을 종료할 때 호출합니다.
+	 *
+	 * @param user 사용자 엔티티
+	 */
+	public void clearFcmTokens(User user) {
+		if (user.getFcmTokens() == null || user.getFcmTokens().isEmpty()) {
+			return;
+		}
+		for (String token : user.getFcmTokens()) {
+			redisUtils.deleteFcmTokenData(token);
+		}
+		user.getFcmTokens().clear();
 	}
 }
