+++ app-main/src/main/java/net/causw/app/main/domain/user/account/entity/user/User.java
@@ -145,9 +145,6 @@ public class User extends BaseEntity {
 	public void withdraw(LocalDateTime now) {
 		this.deletedAt = now;
 		this.academicStatus = AcademicStatus.UNDETERMINED;
-		if (this.fcmTokens != null) {
-			this.fcmTokens.clear();
-		}
 	}
 
 	/*
