+++ app-main/src/main/java/net/causw/app/main/domain/user/account/service/UserAccountService.java
@@ -10,6 +10,7 @@ import net.causw.app.main.domain.asset.file.entity.joinEntity.UserProfileImage;
 import net.causw.app.main.domain.asset.file.service.v2.implementation.UserProfileImageReader;
 import net.causw.app.main.domain.asset.locker.service.v2.implementation.LockerReader;
 import net.causw.app.main.domain.asset.locker.service.v2.implementation.LockerWriter;
+import net.causw.app.main.domain.notification.notification.service.implementation.UserPushTokenWriter;
 import net.causw.app.main.domain.user.account.api.v2.dto.response.UserWithdrawResponse;
 import net.causw.app.main.domain.user.account.entity.user.SocialAccount;
 import net.causw.app.main.domain.user.account.entity.user.User;
@@ -19,7 +20,6 @@ import net.causw.app.main.domain.user.account.service.dto.result.UserMeAccountRe
 import net.causw.app.main.domain.user.account.service.dto.result.UserMeResult;
 import net.causw.app.main.domain.user.account.service.implementation.SocialAccountReader;
 import net.causw.app.main.domain.user.account.service.implementation.SocialAccountUnlinkManager;
-import net.causw.app.main.domain.user.account.service.implementation.UserInfoReader;
 import net.causw.app.main.domain.user.account.service.implementation.UserReader;
 import net.causw.app.main.domain.user.account.service.implementation.UserValidator;
 import net.causw.app.main.domain.user.account.service.implementation.UserWriter;
@@ -36,7 +36,6 @@ import net.causw.app.main.domain.user.terms.service.implementation.UserTermsAgre
 import net.causw.app.main.shared.dto.ProfileImageDto;
 import net.causw.app.main.shared.exception.errorcode.AuthErrorCode;
 import net.causw.app.main.shared.exception.errorcode.UserErrorCode;
-import net.causw.app.main.shared.infra.firebase.FcmUtils;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -55,14 +54,13 @@ public class UserAccountService {
 	private final UserValidator userValidator;
 	private final AuthValidator authValidator;
 	private final AuthTokenManager authTokenManager;
-	private final FcmUtils fcmUtils;
 	private final PasswordEncoder passwordEncoder;
 	private final UserProfileImageReader userProfileImageReader;
 	private final TermsReader termsReader;
 	private final TermsValidator termsValidator;
 	private final UserTermsAgreementReader userTermsAgreementReader;
 	private final UserTermsAgreementWriter userTermsAgreementWriter;
-	private final UserInfoReader userInfoReader;
+	private final UserPushTokenWriter userPushTokenWriter;
 	private final UserProfileImageService userProfileImageService;
 
 	/**
@@ -314,8 +312,8 @@ public class UserAccountService {
 		// 부가 처리
 		lockerReader.findByUserId(user.getId())
 			.ifPresent(locker -> lockerWriter.returnLocker(locker, user));
-		fcmUtils.clearFcmTokens(user);
 
+		userPushTokenWriter.clearFcmTokens(user);
 		userWriter.withdraw(user);
 
 		return UserWithdrawResponse.of(user.getDeletedAt());
