+++ app-main/src/test/java/net/causw/app/main/domain/user/account/service/UserAccountServiceTest.java
@@ -29,6 +29,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import net.causw.app.main.domain.asset.locker.entity.Locker;
 import net.causw.app.main.domain.asset.locker.service.v2.implementation.LockerReader;
 import net.causw.app.main.domain.asset.locker.service.v2.implementation.LockerWriter;
+import net.causw.app.main.domain.notification.notification.service.implementation.UserPushTokenWriter;
 import net.causw.app.main.domain.user.account.api.v2.dto.response.UserWithdrawResponse;
 import net.causw.app.main.domain.user.account.entity.user.SocialAccount;
 import net.causw.app.main.domain.user.account.entity.user.User;
@@ -50,7 +51,6 @@ import net.causw.app.main.domain.user.terms.service.implementation.TermsValidato
 import net.causw.app.main.domain.user.terms.service.implementation.UserTermsAgreementWriter;
 import net.causw.app.main.shared.exception.BaseRunTimeV2Exception;
 import net.causw.app.main.shared.exception.errorcode.TermsErrorCode;
-import net.causw.app.main.shared.infra.firebase.FcmUtils;
 
 @ExtendWith(MockitoExtension.class)
 class UserAccountServiceTest {
@@ -104,7 +104,7 @@ class UserAccountServiceTest {
 	private LockerWriter lockerWriter;
 
 	@Mock
-	private FcmUtils fcmUtils;
+	private UserPushTokenWriter userPushTokenWriter;
 
 	@Mock
 	private UserProfileImageService userProfileImageService;
@@ -253,7 +253,7 @@ class UserAccountServiceTest {
 		verify(authTokenManager).invalidateTokens(accessToken, refresh);
 		verify(lockerReader).findByUserId(userId);
 		verify(lockerWriter).returnLocker(locker, user);
-		verify(fcmUtils).clearFcmTokens(user);
+		verify(userPushTokenWriter).clearFcmTokens(user);
 
 		verify(userWriter).withdraw(user);
 	}
@@ -299,6 +299,7 @@ class UserAccountServiceTest {
 			() -> userAccountService.withdraw(userId, "access-token", "refresh-token"));
 
 		verify(userReader).findUserById(userId);
-		verifyNoInteractions(socialAccountReader, socialAccountUnlinkManager, lockerReader, lockerWriter, fcmUtils);
+		verifyNoInteractions(socialAccountReader, socialAccountUnlinkManager, lockerReader, lockerWriter,
+			userPushTokenWriter);
 	}
 }
